+++ HISTORY.md
@@ -12,6 +12,9 @@
 * **Translation Updates:**
     * German (for error messages)
 
+### 2022-01-21
+* threads: refactored Context >> image
+
 ### 2022-01-20
 * German translation update (for error messages)
 * threads: added more localizable error strings